Move Intake outside of instance
